requirements.txt
colorama==0.4.6
termcolor==2.4.0
udapi @ git+https://github.com/udapi/udapi-python.git@a43cd21f962a72e4f5e6f5dc40c5d35393669107
ufal.udpipe==1.3.1.1
